docs(client): fix JSDoc comment
The type definitions from the "Product Entry" feature file were not able
to resolve in the "Product Grid" feature file. This was originally
caused by a missing export statement, which has been fixed in a previous
change. At this point, only an import() statement is needed to eliminate
the issue

Changes:
- Add missing import() statement in a JSDoc comment for references to
the "Product Entry" feature file